功能修改：look_back_cnn入库结果由0和1改为float[0,1]。

+++ gwac-core/src/main/java/com/gwac/model/OtLevel2.java
@@ -55,7 +55,7 @@ public class OtLevel2  implements java.io.Serializable {
      private Short usnoMatch;
      private Short lookBackResult;
      private Short followUpResult;
-     private Short lookBackCnn;
+     private Float lookBackCnn;
 
     public OtLevel2() {
     }
@@ -444,14 +444,14 @@ public class OtLevel2  implements java.io.Serializable {
    * @return the lookBackCnn
    */
   @Column(name="look_back_cnn")
-  public Short getLookBackCnn() {
+  public Float getLookBackCnn() {
     return lookBackCnn;
   }
 
   /**
    * @param lookBackCnn the lookBackCnn to set
    */
-  public void setLookBackCnn(Short lookBackCnn) {
+  public void setLookBackCnn(Float lookBackCnn) {
     this.lookBackCnn = lookBackCnn;
   }
